fs_cli: fix 32 bits compile
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@15966 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ libs/esl/fs_cli.c
@@ -437,7 +437,7 @@ static unsigned char complete(EditLine * el, int ch)
 	unsigned char ret = CC_REDISPLAY;
 
 	if (*lf->cursor) {
-		snprintf(cmd_str, sizeof(cmd_str), "api console_complete c=%ld;%s\n\n", lf->cursor - lf->buffer, lf->buffer);
+		snprintf(cmd_str, sizeof(cmd_str), "api console_complete c=%ld;%s\n\n", (long int)(lf->cursor - lf->buffer), lf->buffer);
 	} else {
 		snprintf(cmd_str, sizeof(cmd_str), "api console_complete %s\n\n", lf->buffer);
 	}
